pyproject.toml
[project]
name = "20250125-ai-x-finance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "gpt-researcher>=0.10.11",
    "langchain>=0.3.15",
    "langchain-openai>=0.3.2",
    "loguru>=0.7.3",
    "openai>=1.60.1",
    "pandas>=2.2.3",
    "praw>=7.8.1",
    "pypst>=0.6.1",
    "python-dotenv>=1.0.1",
    "yfinance>=0.2.52",
]

[dependency-groups]
dev = [
    "ipython>=8.31.0",
    "jupyter>=1.1.1",
    "jupyterlab>=4.3.4",
    "matplotlib>=3.10.0",
    "mypy>=1.14.1",
    "notebook>=7.3.2",
    "pandas-stubs>=2.2.3.241126",
    "ruff>=0.9.3",
]
